commit messes all fixed, half precision removed

+++ include/nps_uw_sensors_gazebo/gazebo_ros_image_sonar.hh
@@ -61,7 +61,7 @@
 
 namespace gazebo
 {
-  typedef std::complex<int> Complex;
+  typedef std::complex<float> Complex;
   typedef std::valarray<Complex> CArray;
   typedef std::valarray<CArray> CArray2D;
